Moves Charts Livewire components into a dedicated folder
